+++ tests/api/test_base.py
@@ -89,7 +89,7 @@ def test_api_base_get_service_schema(mocker):
 
     svc_name = 'foo'
     svc_version = '1.0.0'
-    path = '{}/{}'.format(svc_name, svc_version)
+    path = '{} {}'.format(svc_name, svc_version)
     payload = {'foo': 'bar'}
 
     # Check error for empty schemas
@@ -98,7 +98,7 @@ def test_api_base_get_service_schema(mocker):
         api.get_service_schema(svc_name, svc_version)
 
     # Check that schema registry get was called with proper service path
-    registry.get.assert_called_with(path, None)
+    registry.get.assert_called_with(path, None, delimiter=' ')
 
     # Check getting a service schema
     init = mocker.patch(
@@ -108,7 +108,7 @@ def test_api_base_get_service_schema(mocker):
     registry.get = mocker.MagicMock(return_value=payload)
     svc_schema = api.get_service_schema(svc_name, svc_version)
     assert isinstance(svc_schema, ServiceSchema)
-    registry.get.assert_called_with(path, None)
+    registry.get.assert_called_with(path, None, delimiter=' ')
     init.assert_called_with(svc_name, svc_version, payload)
 
     # Check getting a service schema using a wildcard version